Move code from 'export_to_galaxy.py' to new 'galaxy' library module.
